fixed remove friend api

+++ remove_friend.php
@@ -1,40 +1,74 @@
 <?php
 
 header("Access-Control-Allow-Origin: *");
+include("validate.php");
 
 include("db_info.php");
 
-//GET $user id values
-$user1_id =  $_GET[$user1_id];
-$user2_id = $_GET[$user2_id];
-
-// CHECK IF ALREADY FRIENDS
-$query = $mysqli->prepare("SELECT status FROM relationship WHERE ((user1_id = ? AND user2_id = ? AND status = 'friend') OR (user1_id = ? AND user2_id = ? AND status = 'friend'))");
-$query->bind_param("iiii", $user1_id, $user2_id, $user2_id, $user1_id);
-$query->execute();
-$query->store_result();
-$num_rows = $query->num_rows;
-
 $array_response = [];
 
-if($num_rows == 0){
-  $array_response["error"] = "Invalid request.";
-  $json_response = json_encode($array_response);
+// JWT VALIDATION
 
-  return false;
+if(isset($_POST["token"]) ){
+    
+  $token = $_POST["token"];
+
+  $validate= is_jwt_valid($_POST["token"] ,$secret = 'secret');
+
+  if ($validate){ 
+      $tokenParts = explode('.', $token);
+      $payload = base64_decode($tokenParts[1]);
+      $data = json_decode($payload, true);
+
+      $user1_id= $data["id"];
+
+      if(isset($_POST["friendID"])){
+        $user2_id = $_POST["friendID"];
+
+        // CHECK IF ALREADY FRIENDS
+        $query = $mysqli->prepare("SELECT status FROM relationship WHERE ((user1_id = ? AND user2_id = ? AND status = 'friend') OR (user1_id = ? AND user2_id = ? AND status = 'friend'))");
+        $query->bind_param("iiii", $user1_id, $user2_id, $user2_id, $user1_id);
+        $query->execute();
+        $query->store_result();
+        $num_rows = $query->num_rows;
+
+        if($num_rows == 0){
+          $array_response["error"] = "Invalid request.";
+          $json_response = json_encode($array_response);
+          echo $json_response;
+          return false;
+        }else{
+            //REMOVE FRIEND - DELETE RELATIONSHIP ENTRY
+            $query = $mysqli->prepare(
+                "DELETE FROM relationship 
+                WHERE ((user1_id =? AND user2_id =? AND status='friend') OR 
+                (user1_id =? AND user2_id =? AND status='friend'))");
+        
+            $query->bind_param("iiii", $user1_id, $user2_id, $user2_id, $user1_id);
+            $query->execute();
+            $array_response["status"] = "Success";
+            $json_response = json_encode($array_response);
+            echo $json_response;
+        }
+      }else{
+        $array_response["error"] = "Missing friendID";
+        $json_response = json_encode($array_response);
+        echo $json_response;
+        return false;
+      }
+  }else{
+      $array_response["error"] = "Invalid token.";
+      $json_response = json_encode($array_response);
+      echo $json_response;
+      return false;
+    }
 }else{
-    //REMOVE FRIEND - DELETE RELATIONSHIP ENTRY
-    global $mysqli;
-    $query = $mysqli->prepare(
-        "DELETE FROM relationship 
-        WHERE ((user1_id =? AND user2_id =? AND status='friend') OR 
-        (user1_id =? AND user2_id =? AND status='friend'))");
-
-    $query->bind_param("iiii", $user1_id, $user2_id, $user2_id, $user1_id);
-    $query->execute();
-    $array_response["status"] = "Friend removed successfully.";
-    $json_response = json_encode($array_response);
-}
+  $array_response["error"] = "Token not received.";
+  $json_response = json_encode($array_response);
+  echo $json_response;
+  return false;
+};
+        
 $query->close();
 $mysqli->close();
 ?>
